Refactor OBMysqlCallFunctionCallBack and move tests
